test: group serialized in-use binary tests into `foundry_bins` module
